Specify dependency on "gakubuchi/configuration"
