+++ lib/gakubuchi.rb
@@ -10,21 +10,3 @@ if defined?(::Rails::Railtie) && defined?(::Sprockets::Railtie)
   require "gakubuchi/railtie"
   require "gakubuchi/template"
 end
-
-module Gakubuchi
-  class << self
-    attr_writer :configuration
-
-    def configuration
-      @configuration ||= ::Gakubuchi::Configuration.new
-    end
-
-    def configure
-      yield(configuration) if block_given?
-    end
-
-    def reset
-      @configuration = nil
-    end
-  end
-end
