+++ lib/gakubuchi/configuration.rb
@@ -2,6 +2,22 @@ require "active_support/configurable"
 require "forwardable"
 
 module Gakubuchi
+  class << self
+    attr_writer :configuration
+
+    def configuration
+      @configuration ||= ::Gakubuchi::Configuration.new
+    end
+
+    def configure
+      yield(configuration) if block_given?
+    end
+
+    def reset
+      @configuration = nil
+    end
+  end
+
   class Configuration
     extend ::Forwardable
     include ::ActiveSupport::Configurable
